subprojects: jris-core, jris-io, jris-guide (#16)
* Split JRis into subprojects jris-core, jris-io and jris-guide
* Bump rxjava2 from 2.2.15 to 2.2.16
* Use java1.8 way of creating a list
