+++ buildSrc/src/main/kotlin/Lib.kt
@@ -9,7 +9,7 @@ object Lib {
     fun kotlin(module: String) = Dep("org.jetbrains.kotlin:kotlin-$module", kotlinVersion)
     fun kotlinx(module: String) = Dep("org.jetbrains.kotlinx:kotlinx-$module", "1.3.3")
 
-    fun rxjava2() = Dep("io.reactivex.rxjava2:rxjava", "2.2.15")
+    fun rxjava2() = Dep("io.reactivex.rxjava2:rxjava", "2.2.16")
     fun rxkotlin() = Dep("io.reactivex.rxjava2:rxkotlin", "2.4.0")
 
     fun junit5(module: String = "") =
