+++ subprojects/jris-core/jris-core.gradle.kts
@@ -1,11 +1,4 @@
-plugins {
-    id("org.kordamp.gradle.integration-test")
-}
-
 dependencies {
     implementation(Lib.rxjava2())
-    implementation(Lib.rxkotlin())
     implementation(Lib.kotlinx("coroutines-rx2"))
-
-    testImplementation(Lib.assertJ())
 }
