+++ subprojects/jris-core/src/main/kotlin/com/gmail/gcolaianni5/jris/BridgeExtensions.kt
@@ -1,3 +1,5 @@
+@file:Suppress("SpellCheckingInspection")
+
 package com.gmail.gcolaianni5.jris
 
 import io.reactivex.Observable
