+++ subprojects/jris-guide/jris-guide.gradle.kts
@@ -1,0 +1,5 @@
+dependencies {
+    testImplementation(project(":jris-io"))
+
+    testImplementation(Lib.rxjava2())
+}
