+++ subprojects/jris-io/jris-io.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("org.kordamp.gradle.integration-test")
+}
+
+dependencies {
+    api(project(":jris-core"))
+}
